cleanup: use unsigned int type as returned by MFloatPointArray.length()

+++ src/serlio/prtModifier/PRTMesh.cpp
@@ -39,7 +39,7 @@ PRTMesh::PRTMesh(const MObject& mesh) {
 	MFloatPointArray vertexArray;
 	meshFn.getPoints(vertexArray);
 
-	const size_t vertexArrayLength = vertexArray.length();
+	const unsigned int vertexArrayLength = vertexArray.length();
 	mVertexCoordsVec.reserve(vertexArrayLength);
 	for (unsigned int i = 0; i < vertexArrayLength; ++i) {
 		mVertexCoordsVec.push_back(vertexArray[i].x);
